Adjust size and position of line and toolbar
UIToolbar seems not to respect the height resize option. Thus, it needs
to be done by hand.

The line is initially placed at a half point from the top. When the
super view height changes, the line is placed on an integer point. Thus,
the line is placed at a half point when laying out the subviews.